Remove XRDeviceImpl
The WebXR API used to have an intermediate device that must be obtained
prior to requesting sessions.  This device was removed from the API.
This change removes the implementation backing it (across a mojo
boundary) in the browser process.

Effectively this change merges XRDeviceImpl and VRServiceImpl and
simplifies some parts of the code related where multiple async
operations were required.

Bug: 944987
Change-Id: If1c426baab8e3149059f32773e9a21313769df0b